Update docs (#197)
* updating build content

* fix link

* fix video

* typos

* tweaking video

* video updates

* media updates

* update principles formatting

* typos

* tweaks

* Update Principles.md

* changing video markup

* Update Principles.md (#64)

* remove media from toc for now

* update roadmap

* update channel status and added link to scuba bot source

* update Teams to be in preview

* Update Partners.md

* push live (#69) (#70)

* CI Update (#71)

* Small correction on element rendering support in HTML renderer (#51)

* fix sample url

* Updates to UWP functionality

* Add a few links

* Fix links for OP

* Restructuring SDK info

* Fix date/time link

* Alt layout

* Alt2

* Rearranging folder structures

* Fix TOC

* Fix nesting toc

* Try removing leading slashes

* Add all redirects

* Fix redirects

* Fix TOC

* Make resources files lowercase

* Sidebar updates

* Fix links

* Fix links

* Fix images

* Fix remaining links

* Update some https links

* Fix Android maven status link

* Update getting starteds

* Update title

* Update titles of pages

* Update UWP action docs

* Update schema version/fallback

* Update renderer versioning

* Remove Xamarin for now

* Add action renderer extensibility row

* Update action renderer extensibility status

Android is supported in master, but not in nuget packages yet, so leaving that as an X for now.

* Update render-a-card.md (#84)

incorrect code, incorrectly spelt variable name

* updated action title to be required

* Update ShowCard action event handler info

Added clarification that ShowCard event only gets fired if HostConfig is set to popup rather than inline

* Update host config and card schema documentation

* updated teams as officially launched

* add version column

* update urls

* Update docfx.json (#114)

Fixing metadata. Please contact @kcpitt for questions

* Spelling fix

Changed to IList<AdaptiveWarning> warnings = renderedCard.Warnings;

* Fixing bad link in GettingStarted.md

We removed the version folder from the hostConfigs

* Update readme (#110)

Update Note to reflect Teams now supporting Adaptive Cards: https://techcommunity.microsoft.com/t5/Microsoft-Teams-Blog/What-s-new-in-Teams-Microsoft-Ignite-Edition/ba-p/252531

* Update bots.md

* Update native-styling.md (#90)

* Update renderer-status.md

* added note that HTTPS and Media aren't working in WPF

* Live (#135)

* push live (#126)

* Push live (#134)

* Fix Outlook Actionable Messages link

The Outlook Actionable Messages link was simply going to outlook.com, which was super confusing... That meant clicking it simply opened my own email inbox!

* Update future.md

* Update extensibility.md (#148)

* Update extensibility.md

* Update extensibility.md

* Update extensibility.md

* Update windows.md (#142)

fixed ms.topic metadata to be compliant.

* Update getting-started.md (#141)

fixed ms.topic metadata to be compliant.

* Fixed broken link to sample HostConfig.json (#139)

* The format accepted for dates and time is strict on the RFC 3389 spec. (#138)

While the Valid formats section is correct in specifying the seconds. The sample card was invalid.

* Update getting-started.md

* Added designer to tools. Closes #155

* tools cleanup

* added note to find sdk extensbility docs

Fixes #160

* fix grammar issue (#158)

* Small typo fix (#173)

Just a small fix:  custom `Rating` element to you define. ->  custom `Rating` element that you define.

* Fixed some typos (#175)

* Updated misspelled variables and comment (#176)

* Add extensibility changes for all versions (#170)

Add extensibility changes for image loading, media loading and an empty section for action loading to be completed when the full version of v1.2 is released

* Pass the AbsoluteUri to Process.Start rather than the URI (#161)

Pass a string to Process.Start so that it compiles

* Update host-config.md (#164)

* Add setter to Type to make custom type not abstract (#165)

* Update sample to use constuctor that takes a version (#163)

The no-argument version of the constructor is deprecated. Updated the sample to use the overload that takes the card version.

* Fix type of Action args (#162)

Update the type from to ActionEventArgs to AdaptiveActionEventArgs

* Add installation and import information (#166)

Adds the information needed to find the maven packages and to import the adaptivecards object model and renderer

* Add styling added for v1.2 (#169)

Add instructions on how to style actions based on sentiments and inline actions, both properties added in v1.2 of AdaptiveCards

* Add actions breaking change from v1.1 (#168)

Add the breaking change from version 1.1 which include the onMediaPlay and onMediaStop actions

* Add information on card rendering for android (#167)

Adds the breaking changes included in v1.2 of adaptive cards, also fixed the original way of rendering the card

* updated doc explaining ios redendering (#189)

* A little cleanup

* FIx Android MediaDataSource link

* Bot Framework updated to 1.2